conf: Add skeleton machine config based on qemux86-64.
This is where we will enable the MACHINE_FEATURE to pull in the tpm-crb
kernel modules. The yocto-linux recipe needs to know about our machine
type, or we need to provide a full kernel config. To hack around this we
set the KMACHINE to 'qemux86-64' effectively lying to the recipe about
the machine. We won't deviate far enough from the qemux86-64 machine for
this to matter.

Signed-off-by: Philip Tricca <flihp@twobit.us>