packagegroup-base: Add bbappend to add the tpm-crb kernel module.
This bbappend follows the general form of the packagegroup-base recipe
from openembedded-core. It creates a packagegroup that does nothing more
than RRECOMMENDS the recipe for the kernel package. We add this new
package to the PROVIDES the RDEPENDS for the packagegroup-base recipe &
package respectively.

All of this allows us to pull the tpm-crb kernel module into an image so
long as the image depends on packagegroup-base and the machine lists
tpm-crb in its MACHINE_FEATURES.

Signed-off-by: Philip Tricca <flihp@twobit.us>